Fix bad usages of mktime reported by PHPstan

+++ front/stat.item.php
@@ -51,8 +51,8 @@ if (isset($_GET["date2"])) {
 }
 
 if (empty($_POST["date1"]) && empty($_POST["date2"])) {
-    $year           = date("Y") - 1;
-    $_POST["date1"] = date("Y-m-d", mktime(1, 0, 0, date("m"), date("d"), $year));
+    $year           = ((int) date("Y")) - 1;
+    $_POST["date1"] = date("Y-m-d", mktime(1, 0, 0, (int) date("m"), (int) date("d"), $year));
     $_POST["date2"] = date("Y-m-d");
 }
